feat: use pickle tmp  python variable

+++ src/_data/py-tool/_finder.py
@@ -1,4 +1,4 @@
-import os
+import pickle,os
 # credit: https://robvanderg.github.io/scripts/scripts/
 class ScriptFinder():
     def __init__(self):
@@ -7,28 +7,35 @@ class ScriptFinder():
         downloads them to a text file, and loads them to a list, where every index
         of valid unicode is represented by a string that contains the script name.
         Note that this is not very RAM efficient, but very fast for lookups.
-        """
-        self.ranges = [None] * 0x110000
-        if not os.path.isfile('static/Scripts.txt'):
-            os.system('mkdir -p static')
-            os.system('wget https://www.unicode.org/Public/16.0.0/ucd/Scripts.txt --no-check-certificate -O static/Scripts.txt')
-        for line in open('static/Scripts.txt'):
-            tok = line.split(';')
-            # 對 scripts.txt 做一點字串處理，拿出 code block 的定義
-            if line[0]!='#' and len(tok) == 2:
-                char_range_hex = tok[0].strip().split('..')
-                char_range_int = [int(x, 16) for x in char_range_hex]
-                script_name = tok[1].strip().split()[0]
-                if len(char_range_int) == 1:
-                    self.ranges[char_range_int[0]] = script_name
-                    # print(self.ranges[:100])
-                else:
-                    for ind in range(char_range_int[0], char_range_int[1]+1):
-                        self.ranges[ind] = script_name
-                # Note that we include the first and the last character of the
-                # range in the indices, so the first range for Latin is 65-90
-                # for example, character 65 (A) and 90 (Z) are both included in
-                # the Latin set.  
+        """        
+        cache_path = "static/Scripts.pkl"
+        if os.path.isfile(cache_path):
+            with open(cache_path, "rb") as f:
+                self.ranges = pickle.load(f)
+        else:
+            self.ranges = [None] * 0x110000
+            if not os.path.isfile('static/Scripts.txt'):
+                os.system('mkdir -p static')
+                os.system('wget https://www.unicode.org/Public/16.0.0/ucd/Scripts.txt --no-check-certificate -O static/Scripts.txt')
+            for line in open('static/Scripts.txt'):
+                tok = line.split(';')
+                # 對 scripts.txt 做一點字串處理，拿出 code block 的定義
+                if line[0]!='#' and len(tok) == 2:
+                    char_range_hex = tok[0].strip().split('..')
+                    char_range_int = [int(x, 16) for x in char_range_hex]
+                    script_name = tok[1].strip().split()[0]
+                    if len(char_range_int) == 1:
+                        self.ranges[char_range_int[0]] = script_name
+                        # print(self.ranges[:100])
+                    else:
+                        for ind in range(char_range_int[0], char_range_int[1]+1):
+                            self.ranges[ind] = script_name
+                    # Note that we include the first and the last character of the
+                    # range in the indices, so the first range for Latin is 65-90
+                    # for example, character 65 (A) and 90 (Z) are both included in
+                    # the Latin set.  
+            with open(cache_path, "wb") as f:
+                pickle.dump(self.ranges, f)
     def find_char(self, char):
         """
         Return the script of a single character, if a string
